move quirks v2 to their own module
